# src/docs/swagger.yaml
definitions:
  dto.CreatePostDTO:
    properties:
      category:
        type: string
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  dto.UpdatePostDTO:
    properties:
      category:
        type: string
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  models.Post:
    properties:
      category:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /posts:
    get:
      description: Retrieve a list of all posts
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new blog post with the provided details
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePostDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Deleted post successfully
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a post
      tags:
      - posts
    get:
      description: Retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update a post by ID with the provided details
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing post
      tags:
      - posts
  /posts/filter:
    get:
      description: Filter blog posts by a search term
      parameters:
      - description: Search term
        in: query
        name: term
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Posts By Filter
      tags:
      - posts
swagger: "2.0"